refactor: accomplish the migration of pages to 'src/views'
