--- src/components/Content/index.js
@@ -1,3 +1,0 @@
-import Content from './Content.vue';
-
-export default Content;
