+++ src/components/MessageCard/MessageCard.vue
@@ -1,19 +1,27 @@
 <template>
-  <div class="message" :class="{'has-not-read': !hasRead}">
-    <router-link :to="{ name: 'user', params: {loginname: data.author.loginname}}">{{ data.author.loginname }}</router-link>
+  <div class="message"
+       :class="{'has-not-read': !hasRead}">
+    <router-link :to="{
+      name: 'user', params: {loginname: data.author.loginname}
+    }">{{ data.author.loginname }}</router-link>
     回复了你的文章
-    <router-link :to="{ name: 'article', params: {id: data.topic.id}}">{{ data.topic.title }}</router-link>
+    <router-link :to="{
+      name: 'article', params: {id: data.topic.id}
+    }">{{ data.topic.title }}</router-link>
     <p class="reply-time">回复于{{ data.reply.create_at | timeFormat }}</p>
-    <div class="mark" v-if="!hasRead" @click="markOne(data.id)">标记已读</div>
+    <div class="mark"
+         v-if="!hasRead"
+         @click="markOne(data.id)">标记已读</div>
     <div class="content">
       <h2>详细内容：</h2>
-      <div class="markdown-text" v-html="data.reply.content"></div>
+      <div class="markdown-text"
+           v-html="data.reply.content"></div>
     </div>
   </div>
 </template>
 
 <script type="text/ecmascript-6">
-import timeFormat from '../../../../common/utils/timeFormat';
+import timeFormat from '@/common/utils/timeFormat';
 
 export default {
   methods: {
@@ -23,7 +31,7 @@ export default {
           accesstoken: this.accesstoken,
         })
         .then(
-          (res) => {
+          () => {
             this.$store.dispatch('add_success', { content: '标记成功' });
             this.$store.dispatch('mark_one', id);
           },
