+++ src/views/MyCollection/index.js
@@ -1,0 +1,1 @@
+export default from './MyCollection.vue';
